Refactor crime application completion routes
